pyproject.toml
[tool.poetry]
name = "traktor-kontrol-x1-mk1"
version = "0.1.0"
description = ""
authors = ["Vincent Letarouilly <vincent.letarouilly@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyusb = "^1.2.1"
mido = "^1.3.2"
python-rtmidi = "^1.5.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
